structured the files
